+++ eslint.config.mjs
@@ -5,7 +5,7 @@ import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
   {
-    ignores: ['build', 'coverage', 'node_modules'],
+    ignores: ['build', 'coverage', 'node_modules', 'preview/dist'],
   },
   {
     ...js.configs.recommended,